requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
geojson==2.5.0
idna==3.10
importlib_metadata==8.7.0
Pillow==9.5.0
prettytable==3.16.0
pyowm==3.3.0
PySocks==1.7.1
requests==2.32.3
rgbmatrix==0.0.1
urllib3==2.4.0
wcwidth==0.2.13
zipp==3.23.0
